Permission fixes for files Staff member changes

+++ app/Http/Controllers/FileMainController.php
@@ -5,17 +5,27 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\FileMain;
 use App\Http\Controllers\ActivityLogController;
+use Illuminate\Support\Facades\Storage;
 
 class FileMainController extends Controller
 {
     public function show(){
+        $this->authorize('create', FileMain::class);
+
         $files = FileMain::orderBy('created_at', 'desc')->get();
+
         return view('files.show', ['files'=>$files]);
     }
+
     public function create(){
+        $this->authorize('create', FileMain::class);
+
         return view('files.create');
     }
+
     public function store(Request $request){
+        $this->authorize('create', FileMain::class);
+
         $validatedData = $request->validate([
             'file' => 'required|mimes:pdf'
            ]);
@@ -23,13 +33,15 @@ class FileMainController extends Controller
            $name = $request->file('file')->getClientOriginalName();
     
            //$path = $request->file('file')->store('public/files');
-           $request->file('file')->move(public_path('files'), $name);
+           $path = $request->file->storeAs('public/files', $name);
+
+           $path = ltrim($path, 'public');
     
            $save = new FileMain;
     
            $save->uploader_id = auth()->user()->id;
            $save->name = $name;
-           $save->path = '/files/'.$name;
+           $save->path = $path;
            $save->type = $request->type;
            $save->is_available = 1;
            $save->save();
@@ -41,9 +53,15 @@ class FileMainController extends Controller
     }
     public function destroy($id){
         $file = FileMain::where('id', $id)->first();
-        if(file_exists(public_path($file->path))){
-            unlink(public_path($file->path));
+
+        $this->authorize('delete', $file);
+
+        $path = 'public' . $file->path;
+
+        if(Storage::exists($path)){
+            Storage::delete($path);
         }
+
         $file->delete();
